refactor: proxy dependencies.
