+++ src/common/modules/caches/index.ts
@@ -1,5 +1,4 @@
-import { CacheModule, forwardRef, Module } from '@nestjs/common';
-import { ProcessEnvModule } from '../..';
+import { CacheModule, Module } from '@nestjs/common';
 import { CachesService } from './caches.service';
 import * as redisStore from 'cache-manager-redis-store';
 
@@ -7,7 +6,6 @@ require('dotenv').config(); // eslint-disable-line
 
 @Module({
   imports: [
-    ProcessEnvModule,
     CacheModule.register({
       store: redisStore,
       host: process.env.HOST_REDIS,
