+++ src/common/modules/debio-conversion/debio-conversion.module.ts
@@ -1,10 +1,10 @@
-import { forwardRef, Module } from '@nestjs/common';
-import { ProcessEnvModule } from '../..';
+import { Module } from '@nestjs/common';
+import { ProcessEnvModule } from '../proxies';
 import { DebioConversionService } from './debio-conversion.service';
 
 @Module({
   imports: [ProcessEnvModule],
   providers: [DebioConversionService],
-  exports: [DebioConversionService],
+  exports: [ProcessEnvModule, DebioConversionService],
 })
 export class DebioConversionModule {}
