+++ src/common/modules/substrate/index.ts
@@ -1,11 +1,11 @@
-import { forwardRef, Module } from '@nestjs/common';
-import { ProcessEnvModule } from '../..';
+import { Module } from '@nestjs/common';
+import { ProcessEnvModule } from '../proxies';
 import { SubstrateService } from './substrate.service';
 
 @Module({
   imports: [ProcessEnvModule],
   providers: [SubstrateService],
-  exports: [SubstrateService],
+  exports: [ProcessEnvModule,SubstrateService],
 })
 export class SubstrateModule {}
