+++ src/endpoints/substrate-endpoint/substrate-endpoint.module.ts
@@ -6,7 +6,7 @@ import { LabController } from './labs/lab.controller';
 import { LabService } from './services/lab.service';
 import { OrderController } from './orders/order.controller';
 import { OrderService } from './services/order.service';
-import { ProcessEnvModule, RewardModule, SubstrateModule } from '../../common';
+import { RewardModule, SubstrateModule } from '../../common';
 import { SubstrateController } from './substrate-endpoint.controller';
 
 @Module({
@@ -22,7 +22,6 @@ import { SubstrateController } from './substrate-endpoint.controller';
     }),
     SubstrateModule,
     RewardModule,
-    ProcessEnvModule,
   ],
   exports: [ElasticsearchModule],
   controllers: [
